refactor: Lotto, DefaultLotto, WinningLotto를 별도의 파일로 분리
