+++ Lotto/Lotto/Lotto.swift
@@ -16,17 +16,6 @@ enum LottoConstant {
 
 protocol Lotto {
     var numbers: Set<Int> { get }
-    func equalNumberCount(with lotto: Lotto) -> Int
-}
-
-extension Lotto {
-    func equalNumberCount(with lotto: Lotto) -> Int {
-        let count = lotto.numbers.filter {
-            self.numbers.contains($0)
-        }.count
-        
-        return count
-    }
 }
 
 enum LottoError: Error {
@@ -45,52 +34,12 @@ enum LottoError: Error {
     }
 }
 
-struct WinningLotto: Lotto {
-    var numbers: Set<Int>
-    let bonusNumber: Int
-    
-    init<LottoNumbers: Collection>(numbers: LottoNumbers, bonusNumber: Int) throws where LottoNumbers.Element == Int {
-        let setNumbers = Set(numbers)
+extension Lotto {
+    func equalNumberCount(with lotto: Lotto) -> Int {
+        let count = lotto.numbers.filter {
+            self.numbers.contains($0)
+        }.count
         
-        guard setNumbers.count == LottoConstant.numberCount else {
-            throw LottoError.invalidNumberCount(numbers.count)
-        }
-        
-        let hasValidNumbers = numbers.reduce(true) { partialResult, number in
-            partialResult && LottoConstant.numberRange.contains(number)
-        }
-        if hasValidNumbers == false {
-            throw LottoError.invalidLottoNumber
-        }
-    
-        if bonusNumber < 0 { throw LottoError.invalidLottoNumber }
-        self.numbers = setNumbers
-        self.bonusNumber = bonusNumber
-    }
-    
-    func checkBonusNumber(in lotto: Lotto) -> Bool {
-        let count = lotto.numbers.filter { $0 == bonusNumber }.count
-        return count > 0
-    }
-}
-
-struct UserLotto: Lotto {
-    let numbers: Set<Int>
-    
-    init<LottoNumbers: Collection>(numbers: LottoNumbers) throws where LottoNumbers.Element == Int  {
-        let setNumbers = Set(numbers)
-        
-        guard setNumbers.count == LottoConstant.numberCount else {
-            throw LottoError.invalidNumberCount(numbers.count)
-        }
-        
-        let hasValidNumbers = numbers.reduce(true) { partialResult, number in
-            partialResult && LottoConstant.numberRange.contains(number)
-        }
-        if hasValidNumbers == false {
-            throw LottoError.invalidLottoNumber
-        }
-    
-        self.numbers = setNumbers
+        return count
     }
 }
